+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Internal config management now far simpler, makes changing or adding configurable features much easier
 * pytoil config CLI drastically simplified
 * NEW: Added `pytoil docs` command to open the package documentation in users default browser
+* NEW: Added `pytoil gh` command to open the GitHub page for a project.
 
 ## 0.2.0